requirements-poc.txt
# POC: Kafka ingestion bridge and trade producer
kafka-python==2.0.2
